feat: Migrate CareBearStare to Streamer.bot
This commit migrates the CareBearStare script from the old Streamlabs
Chatbot (IronPython) to the new Streamer.bot (C#).

The new version of the script is located in the CareBearStare-StreamerBot
directory. It includes the C# script, an HTML overlay, and a README.md
file with installation and configuration instructions.

The script has been refactored to use the built-in Streamer.bot methods
for Twitch API calls, removing the need for a clientid.conf file.

A plan.md file has also been added to the root of the repository,
outlining the migration process.